+++ mobile/android/geckoview/src/androidTest/assets/web_extensions/tabs/background.js
@@ -1,0 +1,2 @@
+browser.tabs.create({ url: "https://www.mozilla.org/en-US/" });
+